Dan/atc readout timing (#15)
* Add files for testing downstream timing

These files serve to test and then plot the timing jitter on the downstream end of the ZMQ socket with images from camerad.

* Update and rename TimingJitter_runner.cpp to TimingJitter_DownstreamRunner.cpp

Update filename to make clear which element is being tested.

* Update and rename TimingJitter_plotter.py to TimingJitter_DownstreamPlotter.py

Rename file to make clear which element's timing is being plotted.

* Create README.md

Add some brief explanation of what each file is for, so that we can remember the basics of how a test is run in the future.